refactor: changing dir and namespace
