+++ app/Http/Controllers/EmployeeController.php
@@ -8,9 +8,9 @@ use App\Http\Resources\AddressNameResource;
 use App\Http\Requests\StoreEmployeeRequest;
 use App\Http\Resources\EmployeeResource;
 use Illuminate\Support\Facades\Redirect;
-use App\Actions\UpdateEmployeeAction;
-use App\Actions\DeleteEmployeeAction;
-use App\Actions\StoreEmployeeAction;
+use App\Actions\Employees\UpdateEmployeeAction;
+use App\Actions\Employees\DeleteEmployeeAction;
+use App\Actions\Employees\StoreEmployeeAction;
 use Illuminate\Http\Request;
 use App\Models\EmployeeRole;
 use App\Models\Employee;
